+++ Cargo.toml
@@ -49,6 +49,7 @@ members = [
 	"client/network/test",
 	"client/offchain",
 	"client/peerset",
+	"client/allocator",
 	"client/proposer-metrics",
 	"client/rpc",
 	"client/rpc-api",
@@ -129,7 +130,6 @@ members = [
 	"frame/uniques",
 	"frame/utility",
 	"frame/vesting",
-	"primitives/allocator",
 	"primitives/api",
 	"primitives/api/proc-macro",
 	"primitives/api/test",
@@ -141,7 +141,6 @@ members = [
 	"primitives/authorship",
 	"primitives/block-builder",
 	"primitives/blockchain",
-	"primitives/chain-spec",
 	"primitives/consensus/aura",
 	"primitives/consensus/babe",
 	"primitives/consensus/common",
