+++ client/allocator/README.md
@@ -1,0 +1,6 @@
+Collection of allocator implementations.
+
+This crate provides the following allocator implementations:
+- A freeing-bump allocator: [`FreeingBumpHeapAllocator`](https://docs.rs/sc-allocator/latest/sc_allocator/struct.FreeingBumpHeapAllocator.html)
+
+License: Apache-2.0
