+++ client/chain-spec/Cargo.toml
@@ -20,7 +20,6 @@ sp-core = { version = "3.0.0", path = "../../primitives/core" }
 serde = { version = "1.0.101", features = ["derive"] }
 serde_json = "1.0.41"
 sp-runtime = { version = "3.0.0", path = "../../primitives/runtime" }
-sp-chain-spec = { version = "3.0.0", path = "../../primitives/chain-spec" }
 sc-telemetry = { version = "3.0.0", path = "../telemetry" }
 codec = { package = "parity-scale-codec", version = "2.0.0" }
 sc-consensus-babe = { version = "0.9.0", path = "../consensus/babe" }
