+++ client/executor/common/Cargo.toml
@@ -19,7 +19,7 @@ pwasm-utils = "0.18.0"
 codec = { package = "parity-scale-codec", version = "2.0.0" }
 wasmi = "0.9.0"
 sp-core = { version = "3.0.0", path = "../../../primitives/core" }
-sp-allocator = { version = "3.0.0", path = "../../../primitives/allocator" }
+sc-allocator = { version = "3.0.0", path = "../../allocator" }
 sp-wasm-interface = { version = "3.0.0", path = "../../../primitives/wasm-interface" }
 sp-maybe-compressed-blob = { version = "3.0.0", path = "../../../primitives/maybe-compressed-blob" }
 sp-serializer = { version = "3.0.0", path = "../../../primitives/serializer" }
