+++ client/executor/wasmtime/Cargo.toml
@@ -23,7 +23,7 @@ sc-executor-common = { version = "0.9.0", path = "../common" }
 sp-wasm-interface = { version = "3.0.0", path = "../../../primitives/wasm-interface" }
 sp-runtime-interface = { version = "3.0.0", path = "../../../primitives/runtime-interface" }
 sp-core = { version = "3.0.0", path = "../../../primitives/core" }
-sp-allocator = { version = "3.0.0", path = "../../../primitives/allocator" }
+sc-allocator = { version = "3.0.0", path = "../../allocator" }
 wasmtime = "0.27.0"
 
 [dev-dependencies]
