+++ client/rpc-api/Cargo.toml
@@ -25,7 +25,7 @@ parking_lot = "0.11.1"
 sp-core = { version = "3.0.0", path = "../../primitives/core" }
 sp-version = { version = "3.0.0", path = "../../primitives/version" }
 sp-runtime = { path = "../../primitives/runtime" , version = "3.0.0"}
-sp-chain-spec = { path = "../../primitives/chain-spec" , version = "3.0.0"}
+sc-chain-spec = { path = "../chain-spec" , version = "3.0.0"}
 serde = { version = "1.0.101", features = ["derive"] }
 serde_json = "1.0.41"
 sp-transaction-pool = { version = "3.0.0", path = "../../primitives/transaction-pool" }
