+++ client/rpc-api/src/system/helpers.rs
@@ -20,7 +20,7 @@
 
 use std::fmt;
 use serde::{Serialize, Deserialize};
-use sp_chain_spec::{Properties, ChainType};
+use sc_chain_spec::{Properties, ChainType};
 
 /// Running node's static details.
 #[derive(Clone, Debug)]
