+++ client/rpc/Cargo.toml
@@ -31,7 +31,7 @@ sp-utils = { version = "3.0.0", path = "../../primitives/utils" }
 sp-rpc = { version = "3.0.0", path = "../../primitives/rpc" }
 sp-keystore = { version = "0.9.0", path = "../../primitives/keystore" }
 sp-state-machine = { version = "0.9.0", path = "../../primitives/state-machine" }
-sp-chain-spec = { version = "3.0.0", path = "../../primitives/chain-spec" }
+sc-chain-spec = { version = "3.0.0", path = "../chain-spec" }
 sc-executor = { version = "0.9.0", path = "../executor" }
 sc-block-builder = { version = "0.9.0", path = "../block-builder" }
 sc-keystore = { version = "3.0.0", path = "../keystore" }
