--- primitives/allocator/README.md
@@ -1,6 +1,0 @@
-Collection of allocator implementations.
-
-This crate provides the following allocator implementations:
-- A freeing-bump allocator: [`FreeingBumpHeapAllocator`](https://docs.rs/sp-allocator/latest/sp_allocator/struct.FreeingBumpHeapAllocator.html)
-
-License: Apache-2.0
