--- primitives/chain-spec/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "sp-chain-spec"
-version = "3.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-description = "Substrate chain configurations types."
-readme = "README.md"
-
-[dependencies]
-serde = { version = "1.0.101", features = ["derive"] }
-serde_json = "1.0.41"
