--- primitives/chain-spec/README.md
@@ -1,3 +1,0 @@
-Types and traits related to chain specifications.
-
-License: Apache-2.0
