--- primitives/chain-spec/src/lib.rs
@@ -1,43 +1,0 @@
-// This file is part of Substrate.
-
-// Copyright (C) 2020-2021 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! Types and traits related to chain specifications.
-
-/// The type of a chain.
-///
-/// This can be used by tools to determine the type of a chain for displaying
-/// additional information or enabling additional features.
-#[derive(serde::Serialize, serde::Deserialize, Debug, PartialEq, Clone)]
-pub enum ChainType {
-	/// A development chain that runs mainly on one node.
-	Development,
-	/// A local chain that runs locally on multiple nodes for testing purposes.
-	Local,
-	/// A live chain.
-	Live,
-	/// Some custom chain type.
-	Custom(String),
-}
-
-impl Default for ChainType {
-	fn default() -> Self {
-		Self::Live
-	}
-}
-
-/// Arbitrary properties defined in chain spec as a JSON object
-pub type Properties = serde_json::map::Map<String, serde_json::Value>;
